% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_raw_file_info.R,
%   R/save_raw_file_info_sum.R
\name{analyze_raw_file_info}
\alias{analyze_raw_file_info}
\alias{save_raw_file_info_sum}
\title{Analyze raw file info}
\usage{
analyze_raw_file_info(rawFileInfo, maxinjectcutoff)

save_raw_file_info_sum(rawFileInfoSum, outputDir, use_tdreport, tdReportName)
}
\arguments{
\item{rawFileInfo}{Dataframe containing all raw file data.}

\item{maxinjectcutoff}{Fraction of max injection time that should be used as cutoff for max injections.}

\item{rawFileInfoSum}{Dataframe containing all raw file summary data.}

\item{outputDir}{Path to directory for output of summary info sheet.}

\item{use_tdreport}{Boolean value. Should raw file names be retrieved from a tdReport?}

\item{tdReportName}{Name of the tdReport file.}
}
\value{
Dataframe containing analysis of raw file info.

Saves XLSX to output dir containing raw file info summary.
}
\description{
Analyze raw file info

Save raw file info sum
}
\examples{
Fix this later
}
