% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawfish.R
\name{crawfish}
\alias{crawfish}
\title{Create Raw File Information Sheet}
\usage{
crawfish(
  rawFileDir = NULL,
  use_tdreport = FALSE,
  tdReportDir = NULL,
  tdReportName = NULL,
  outputDir = NULL,
  maxinjectcutoff = 0.99,
  make_report = TRUE
)
}
\arguments{
\item{rawFileDir}{Full path to the directory containing relevant raw files.}

\item{use_tdreport}{Boolean value. Should raw file names be retrieved from a tdReport?}

\item{tdReportDir}{Full path to the directory containing the tdReport.}

\item{tdReportName}{Name of the tdReport file.}

\item{maxinjectcutoff}{Fraction of max injection time that should be used as cutoff for max injections.}

\item{make_report}{Boolean value. Create an HTML report in the output directory?}
}
\value{
This function saves a table of summary data (.xlsx) and an html report for the relevant raw files to the specified output directory.
}
\description{
Create Raw File Information Sheet
}
\examples{
IOU some examples. -David
}
