% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_file_info_Directory.R
\name{get_raw_file_info_Directory}
\alias{get_raw_file_info_Directory}
\title{Get raw file info, Directory}
\usage{
get_raw_file_info_Directory(rawFileDir)
}
\arguments{
\item{rawFileDir}{Full path to the directory containing relevant raw files.}
}
\value{
Dataframe containing all raw info from all raw files in the input directory.
}
\description{
Get raw file info, Directory
}
\examples{
Fix this later
}
