% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_file_info_TDReport.R
\name{get_raw_file_info_TDReport}
\alias{get_raw_file_info_TDReport}
\title{Get raw file info, TDReport}
\usage{
get_raw_file_info_TDReport(rawFileDir, tdReportDir, tdReportName)
}
\arguments{
\item{rawFileDir}{Full path to the directory containing relevant raw files.}

\item{tdReportDir}{Full path to the directory containing the tdReport.}

\item{tdReportName}{Name of the tdReport file.}
}
\value{
Dataframe containing all raw info from all raw files in the tdReport.
}
\description{
Get raw file info, TDReport
}
\examples{
Fix this later
}
