% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_file_list.R
\name{get_raw_file_list}
\alias{get_raw_file_list}
\title{get_raw_file_list}
\usage{
get_raw_file_list(
  rawFileDir = NULL,
  tdReportDir = NULL,
  tdReportName = NULL,
  use_tdreport = NULL
)
}
\arguments{
\item{rawFileDir}{Full path to the directory containing relevant raw files.}

\item{tdReportDir}{Full path to the directory containing the tdReport.}

\item{tdReportName}{Name of the tdReport file.}

\item{use_tdreport}{Boolean value. Should raw file names be retrieved from a tdReport?}
}
\value{
List containing full paths of raw files
}
\description{
get_raw_file_list
}
\examples{
Fix this later
}
