% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_BPC_ms1}
\alias{make_BPC_ms1}
\title{Make BPC MS1}
\usage{
make_BPC_ms1(rawfilename, rawfiledata = NULL, plot_theme = NULL)
}
\arguments{
\item{rawfilename}{String containing the name of a single raw file.}

\item{rawfiledata}{A dataframe containing data extracted from the corresponding raw file.}

\item{plot_theme}{Optional list of ggplot protos to add to plots.}
}
\description{
Make BPC MS1
}
