% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_TIC_ms2}
\alias{make_TIC_ms2}
\title{Make TIC MS2}
\usage{
make_TIC_ms2(rawfilename, rawfiledata = NULL, plot_theme = NULL)
}
\arguments{
\item{rawfilename}{String containing the name of a single raw file.}

\item{rawfiledata}{A dataframe containing data extracted from the corresponding raw file.}

\item{plot_theme}{Optional list of ggplot protos to add to plots.}
}
\description{
Make TIC MS2
}
