% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_tdreport_filenames}
\alias{read_tdreport_filenames}
\title{Read tdReport Filenames}
\usage{
read_tdreport_filenames(tdreport)
}
\arguments{
\item{tdreport}{Full path to a tdreport file.}
}
\description{
Read tdReport Filenames
}
