% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_raw_file_info.R
\name{save_raw_file_info}
\alias{save_raw_file_info}
\title{Save raw file info}
\usage{
save_raw_file_info(rawFileInfo, outputDir, use_tdreport, tdReportName)
}
\arguments{
\item{rawFileInfo}{Dataframe containing all raw file summary data.}

\item{outputDir}{Path to directory for output of summary info sheet.}

\item{use_tdreport}{Boolean value. Should raw file names be retrieved from a tdReport?}

\item{tdReportName}{Name of the tdReport file.}
}
\value{
Saves XLSX to output dir containing raw file info summary.
}
\description{
Save raw file info
}
\examples{
Fix this later
}
